Minor updates
- iconData moved to /lib folder.
- Incorrect logo color fixed on sidebarOpen.
