+++ pages/_app.js
@@ -31,570 +31,6 @@ export const ACTIONS = {
   CLOSE_ELEMENT: "close-element",
 };
 
-// Icon Store
-export const iconData = {
-  core: [
-    {
-      name: "C",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/c-colored.svg",
-      iTag: "c",
-      link: "https://docs.microsoft.com/en-us/cpp/?view=msvc-170",
-    },
-    {
-      name: "C++",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/cplusplus-colored.svg",
-      iTag: "cplusplus",
-      link: "https://docs.microsoft.com/en-us/cpp/?view=msvc-170",
-    },
-    {
-      name: "C#",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/csharp-colored.svg",
-      iTag: "csharp",
-      link: "https://docs.microsoft.com/en-us/dotnet/csharp/",
-    },
-    {
-      name: "Coffeescript",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/coffeescript-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/coffeescript-colored-dark.svg",
-      iTag: "coffeescript",
-      link: "https://coffeescript.org/",
-    },
-    {
-      name: "Dart",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/dart-colored.svg",
-      iTag: "dart",
-      link: "https://dart.dev/",
-    },
-    {
-      name: "Git",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/git-colored.svg",
-      iTag: "git",
-      link: "https://git-scm.com/",
-    },
-    {
-      name: "Go",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/go-colored.svg",
-      iTag: "go",
-      link: "https://go.dev/doc/",
-    },
-    {
-      name: "Java",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/java-colored.svg",
-      iTag: "java",
-      link: "https://www.oracle.com/java/",
-    },
-    {
-      name: "JavaScript",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/javascript-colored.svg",
-      iTag: "javascript",
-      link: "https://developer.mozilla.org/en-US/docs/Web/JavaScript",
-    },
-    {
-      name: "Kotlin",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/kotlin-colored.svg",
-      iTag: "kotlin",
-      link: "https://kotlinlang.org/",
-    },
-    {
-      name: "Perl",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/perl-colored.svg",
-      iTag: "perl",
-      link: "https://www.perl.org/",
-    },
-    {
-      name: "PHP",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/php-colored.svg",
-      iTag: "php",
-      link: "https://www.php.net/",
-    },
-    {
-      name: "Python",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/python-colored.svg",
-      iTag: "python",
-      link: "https://www.python.org/",
-    },
-    {
-      name: "rlang",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/rlang-colored.svg",
-      iTag: "rlang",
-      link: "https://www.r-project.org/",
-    },
-    {
-      name: "Ruby",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/ruby-colored.svg",
-      iTag: "ruby",
-      link: "https://www.ruby-lang.org/en/",
-    },
-    {
-      name: "Rust",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/rust-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/rust-colored-dark.svg",
-      iTag: "rust",
-      link: "https://www.rust-lang.org/",
-    },
-    {
-      name: "Swift",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/swift-colored.svg",
-      iTag: "swift",
-      link: "https://developer.apple.com/swift/",
-    },
-    {
-      name: "TypeScript",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/typescript-colored.svg",
-      iTag: "typescript",
-      link: "https://www.typescriptlang.org/",
-    },
-  ],
-  frontend: [
-    {
-      name: "HTML5",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/html5-colored.svg",
-      iTag: "html5",
-      link: "https://developer.mozilla.org/en-US/docs/Glossary/HTML5",
-    },
-    {
-      name: "React",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/react-colored.svg",
-      iTag: "react",
-      link: "https://reactjs.org/",
-    },
-    {
-      name: "NextJs",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/nextjs-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/nextjs-colored-dark.svg",
-      iTag: "nextjs",
-      link: "https://nextjs.org/docs",
-    },
-    {
-      name: "Vue",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/vuejs-colored.svg",
-      iTag: "vuejs",
-      link: "https://vuejs.org/",
-    },
-    {
-      name: "Nuxtjs",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/nuxtjs-colored.svg",
-      iTag: "nuxtjs",
-      link: "https://nuxtjs.org/",
-    },
-    {
-      name: "Gatsby",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/gatsby-colored.svg",
-      iTag: "gatsby",
-      link: "https://www.gatsbyjs.com/",
-    },
-    {
-      name: "Angular",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/angularjs-colored.svg",
-      iTag: "angularjs",
-      link: "https://angular.io/",
-    },
-    {
-      name: "JQuery",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/jquery-colored.svg",
-      iTag: "jquery",
-      link: "https://jquery.com/",
-    },
-    {
-      name: "CSS3",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/css3-colored.svg",
-      iTag: "css3",
-      link: "https://www.w3.org/TR/CSS/#css",
-    },
-    {
-      name: "Sass",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/sass-colored.svg",
-      iTag: "sass",
-      link: "https://sass-lang.com/",
-    },
-    {
-      name: "TailwindCSS",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/tailwindcss-colored.svg",
-      iTag: "tailwindcss",
-      link: "https://tailwindcss.com/",
-    },
-    {
-      name: "Chakra UI",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/chakra-colored.svg",
-      iTag: "chakra",
-      link: "https://chakra-ui.com/",
-    },
-    {
-      name: "Bootstrap",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/bootstrap-colored.svg",
-      iTag: "bootstrap",
-      link: "https://getbootstrap.com/",
-    },
-    {
-      name: "Material UI",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/materialui-colored.svg",
-      iTag: "materialui",
-      link: "https://mui.com/",
-    },
-    {
-      name: "Redux",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/redux-colored.svg",
-      iTag: "redux",
-      link: "https://redux.js.org/",
-    },
-    {
-      name: "Webpack",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/webpack-colored.svg",
-      iTag: "webpack",
-      link: "https://webpack.js.org/",
-    },
-    {
-      name: "Babel",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/babel-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/babel-colored-dark.svg",
-      iTag: "babel",
-      link: "https://babeljs.io/",
-    },
-    {
-      name: "Svelte",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/svelte-colored.svg",
-      iTag: "svelte",
-      link: "https://svelte.dev/",
-    },
-    {
-      name: "Vite",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/vite-colored.svg",
-      iTag: "vite",
-      link: "https://vitejs.dev/",
-    },
-  ],
-  backend: [
-    {
-      name: "NodeJS",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/nodejs-colored.svg",
-      iTag: "nodejs",
-      link: "https://nodejs.org/en/",
-    },
-    {
-      name: "Express",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/express-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/express-colored-dark.svg",
-      iTag: "express",
-      link: "https://expressjs.com/",
-    },
-    {
-      name: "Fast API",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/fastapi-colored.svg",
-      iTag: "fastapi",
-      link: "https://fastapi.tiangolo.com/",
-    },
-    {
-      name: "GraphQL",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/graphql-colored.svg",
-      iTag: "graphql",
-      link: "https://graphql.org/",
-    },
-    {
-      name: "Oracle",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/oracle-colored.svg",
-      iTag: "oracle",
-      link: "https://www.oracle.com/uk/index.html",
-    },
-    {
-      name: "NestJS",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/nestjs-colored.svg",
-      iTag: "nestjs",
-      link: "https://docs.nestjs.com/",
-    },
-    {
-      name: "MongoDB",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/mongodb-colored.svg",
-      iTag: "mongodb",
-      link: "https://www.mongodb.com/",
-    },
-    {
-      name: "MySQL",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/mysql-colored.svg",
-      iTag: "mysql",
-      link: "https://www.mysql.com/",
-    },
-    {
-      name: "PostgreSQL",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/postgresql-colored.svg",
-      iTag: "postgresql",
-      link: "https://www.postgresql.org/",
-    },
-    {
-      name: "Firebase",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/firebase-colored.svg",
-      iTag: "firebase",
-      link: "https://firebase.google.com/",
-    },
-    {
-      name: "Appwrite",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/appwrite-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/appwrite-colored-dark.svg",
-      iTag: "appwrite",
-      link: "https://appwrite.io/",
-    },
-    {
-      name: "Heroku",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/heroku-colored.svg",
-      iTag: "heroku",
-      link: "https://www.heroku.com/",
-    },
-    {
-      name: "Flask",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/flask-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/flask-colored-dark.svg",
-      iTag: "flask",
-      link: "https://flask.palletsprojects.com/en/2.0.x/",
-    },
-    {
-      name: "Supabase",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/supabase-colored.svg",
-      iTag: "supabase",
-      link: "https://supabase.io/",
-    },
-  ],
-  other: [
-    {
-      name: ".NET",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/dot-net-colored.svg",
-      iTag: "dot-net",
-      link: "https://dotnet.microsoft.com/en-us/",
-    },
-    {
-      name: "Django",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/django-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/django-colored-dark.svg",
-      iTag: "django",
-      link: "https://www.djangoproject.com/",
-    },
-    {
-      name: "Laravel",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/laravel-colored.svg",
-      iTag: "laravel",
-      link: "https://laravel.com/",
-    },
-    {
-      name: "Flutter",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/flutter-colored.svg",
-      iTag: "flutter",
-      link: "https://flutter.dev/",
-    },
-  ],
-  software: [
-    {
-      name: "Photoshop",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/photoshop-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/photoshop-colored-dark.svg",
-      iTag: "photoshop",
-      link: "https://www.adobe.com/uk/products/photoshop.html",
-    },
-    {
-      name: "Illustrator",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/illustrator-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/illustrator-colored-dark.svg",
-      iTag: "illustrator",
-      link: "adobe.com/uk/products/illustrator.html",
-    },
-    {
-      name: "After Effects",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/aftereffects-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/aftereffects-colored-dark.svg",
-      iTag: "aftereffects",
-      link: "https://www.adobe.com/uk/products/aftereffects.html",
-    },
-    {
-      name: "Premiere Pro",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/premierepro-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/premierepro-colored-dark.svg",
-      iTag: "premierepro",
-      link: "https://www.adobe.com/uk/products/premiere.html",
-    },
-    {
-      name: "XD",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/xd-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/xd-colored-dark.svg",
-      iTag: "xd",
-      link: "https://www.adobe.com/uk/products/xd.html",
-    },
-    {
-      name: "Figma",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/figma-colored.svg",
-      iTag: "figma",
-      link: "https://www.figma.com/",
-    },
-    {
-      name: "Sketch",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/sketch-colored.svg",
-      iTag: "sketch",
-      link: "https://www.sketch.com/",
-    },
-  ],
-  web3: [
-    {
-      name: "Uniswap",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/uniswap-colored.svg",
-      iTag: "uniswap",
-      link: "https://uniswap.org/",
-    },
-    {
-      name: "AAVE",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/aave-colored.svg",
-      iTag: "aave",
-      link: "https://aave.com/",
-    },
-    {
-      name: "Sushiswap",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/sushiswap-colored.svg",
-      iTag: "sushiswap",
-      link: "https://www.sushi.com/",
-    },
-    {
-      name: "MetaMask",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/metamask-colored.svg",
-      iTag: "metamask",
-      link: "https://metamask.io/",
-    },
-    {
-      name: "Argent",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/argent-colored.svg",
-      iTag: "argent",
-      link: "https://www.argent.xyz/",
-    },
-    {
-      name: "Nansen",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/nansen-colored.svg",
-      iTag: "nansen",
-      link: "https://www.nansen.ai/",
-    },
-    {
-      name: "Chainlink",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/chainlink-colored.svg",
-      iTag: "chainlink",
-      link: "https://chain.link/",
-    },
-    {
-      name: "The Graph",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/the-graph-colored.svg",
-      iTag: "the-graph",
-      link: "https://thegraph.com/en/",
-    },
-    {
-      name: "Ethers",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/ethers-colored.svg",
-      iTag: "ethers",
-      link: "https://ethers.io",
-    },
-    {
-      name: "Web3Js",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/web3js-colored.svg",
-      iTag: "web3js",
-      link: "https://web3js.readthedocs.io/en/v1.7.1/#",
-    },
-    {
-      name: "Alchemy",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/alchemy-colored.svg",
-      iTag: "alchemy",
-      link: "https://docs.alchemy.com/alchemy/documentation/alchemy-web3",
-    },
-    {
-      name: "Hardhat",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/hardhat-colored.svg",
-      iTag: "hardhat",
-      link: "https://hardhat.org/",
-    },
-    {
-      name: "Truffle",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/truffle-colored.svg",
-      iTag: "truffle",
-      link: "https://trufflesuite.com",
-    },
-    {
-      name: "IPFS",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/ipfs-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/ipfs-colored-dark.svg",
-      iTag: "ipfs",
-      link: "https://ipfs.io/",
-    },
-    {
-      name: "Filebase",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/filebase-colored.svg",
-      iTag: "filebase",
-      link: "https://filebase.com/",
-    },
-    {
-      name: "Arweave",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/arweave-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/arweave-colored-dark.svg",
-      iTag: "arweave",
-      link: "https://www.arweave.org/",
-    },
-    {
-      name: "Ethereum",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/ethereum-colored.svg",
-      iTag: "ethereum",
-      link: "https://ethereum.org/en/",
-    },
-    {
-      name: "Polygon",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/polygon-colored.svg",
-      iTag: "polygon",
-      link: "https://polygon.technology/",
-    },
-    {
-      name: "Arbitrum",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/arbitrum-colored.svg",
-      iTag: "arbitrum",
-      link: "https://portal.arbitrum.one/",
-    },
-    {
-      name: "Avalanche",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/avalanche-colored.svg",
-      iTag: "avalanche",
-      link: "https://www.avax.network/",
-    },
-    {
-      name: "Near",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/near-colored.svg",
-      darkPath:
-        "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/near-colored-dark.svg",
-      iTag: "near",
-      link: "https://near.academy/",
-    },
-    {
-      name: "Flow",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/flow-colored.svg",
-      iTag: "flow",
-      link: "https://www.onflow.org/",
-    },
-    {
-      name: "Solana",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/solana-colored.svg",
-      iTag: "solana",
-      link: "https://solana.com/",
-    },
-    {
-      name: "Terra",
-      path: "https://raw.githubusercontent.com/danielcranney/readme-generator/main/public/icons/skills/terra-colored.svg",
-      iTag: "terra",
-      link: "https://www.terra.money/",
-    },
-  ],
-};
-
 // Color Store
 export const colorStore = {
   lightColors: [
